[feat/#64] 멤버 리뷰 목록 API 주소 변경 및 호출 추가
